requirements.txt
fastapi
uvicorn[standard]
mongoengine
pydantic[email]
python-jose
